+++ .gitignore
@@ -1,4 +1,8 @@
 *.h~
 *.c~
 *.h.gch
+list_etud.csv
+./a.out
+etude
+etud
