--- list_etud.csv
@@ -1,3 +1,0 @@
-ss,dd,2,33.00
-DD,FF,3,33.00
-SS,dd,4,455.00
