+++ src/etud.c
